requirements.txt
fastapi
pydantic
discord.py
uvicorn
yt-dlp
youtube-search-python
